gecode @3.7.3 New port
New Portfile for gecode, a C++ library described by the authors as "a
toolkit for developing constraint-based systems and applications". It is
a dependency of the dep_selector RubyGem being used by Opscode and
others, and a way to install it via MacPorts would be nice.  Default
variant has 1 dependency on boost. Port has 1 variant (+qt) that adds qt
support and depends on qt4-mac.